vtolpathfollower: move loitering -> control layer
